# openapi/openapi.yaml
openapi: 3.0.2
info:
    title: Hourly API
    version: '0.1'
    description: Provides interaction with Hourly web app.
    contact:
        name: Hourly
        url: 'http://www.hourly.com'
        email: test@test.com
servers:
    -
        url: 'http://localhost:8080/api/v0'
        description: Represents the local server.
    -
        url: 'http://hourly-timesheet-dev.herokuapp.com/api/v0'
        description: The staging API environment.
    -
        url: 'http://hourly-timesheet-prod.herokuapp.com/api/v0'
        description: The prod API environment.
paths:
...
    /employees:
        get:
            tags:
                - Employees
            parameters:
                -
                    style: form
                    explode: true
                    name: q
                    description: default query parameter
                    schema:
                        type: string
                    in: query
                    required: false
                -
                    style: form
                    explode: true
                    name: include_totals
                    description: include the total count in the response
                    schema:
                        type: string
                    in: query
                    required: false
                -
                    style: form
                    explode: true
                    name: page
                    description: The page to query from. This will take precedence over a limit and offset.
                    schema:
                        type: string
                    in: query
                    required: false
                -
                    style: form
                    explode: true
                    name: fields
                    description: 'Fields to exclude {"measurements": 0} or only include {"measurements": 1}'
                    schema:
                        type: string
                    in: query
                    required: false
                -
                    style: form
                    explode: true
                    name: sort
                    description: >-
                        comma-separated list of fields to define the sort order. To indicate sorting
                        direction, may be prefixed with + (ascending) or - (descending), e.g.
                        /sales-orders?sort=+id
                    schema:
                        type: string
                    in: query
                    required: false
                -
                    style: form
                    explode: true
                    name: offset
                    description: numeric offset of the first element on a page
                    schema:
                        type: integer
                    in: query
                    required: false
                -
                    style: form
                    explode: true
                    name: limit
                    description: client suggested limit to restrict the number of entries on a page
                    schema:
                        type: integer
                    in: query
                    required: false
                -
                    style: form
                    explode: true
                    name: lang
                    description: 'language code to prefer for document elements (en-US, fr ...) page'
                    schema:
                        type: string
                    in: query
                    required: false
            responses:
                '200':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/UserListResponse'
                    description: Successully retrieved list of users.
                '429':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorListResponse'
                    description: API rate limit exceeded.
                '500':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorListResponse'
                    description: Something is broken.
                '502':
                    description: The cloud is down or being upgraded.
                '503':
                    description: The cloud is up but overloaded with requests.
                '504':
                    description: The cloud is up but there is a failure somewhere causing the request to fail.
            security:
                -
                    bearerAuth: []
            operationId: list_users
            summary: List Employees
            x-openapi-router-controller: domains.employees.controllers.get.controller
        post:
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/User'
                x-body-name: employee
            tags:
                - Employees
            responses:
                '200':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/UserSignUpResponse'
                    description: User registration successful!
                '429':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorListResponse'
                    description: API rate limit exceeded.
                '500':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorListResponse'
                    description: Something is broken.
                '502':
                    description: The cloud is down or being upgraded.
                '503':
                    description: The cloud is up but overloaded with requests.
                '504':
                    description: The cloud is up but there is a failure somewhere causing the request to fail.
            security:
                -
                    bearerAuth: []
            operationId: add_employee
            summary: Add Employee
            x-openapi-router-controller: domains.employees.controllers.post.controller
    /employees/validate:
        post:
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UserValidationList'
                x-body-name: employee_validations
            tags:
                - Employees
            responses:
                '200':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ListResponse'
                    description: User validation successful!
                '429':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorListResponse'
                    description: API rate limit exceeded.
                '500':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorListResponse'
                    description: Something is broken.
                '502':
                    description: The cloud is down or being upgraded.
                '503':
                    description: The cloud is up but overloaded with requests.
                '504':
                    description: The cloud is up but there is a failure somewhere causing the request to fail.
            security:
                -
                    bearerAuth: []
            operationId: validate_employees
            summary: Validate Employees
            x-openapi-router-controller: domains.employees.controllers.post.controller
    '/users/profile/{id}':
        get:
            tags:
                - Employees
            parameters:
                -
                    style: simple
                    explode: false
                    name: id
                    schema:
                        type: string
                    in: path
                    required: true
            responses:
                '200':
                    description: Successfully retrieved Employee profile
                '429':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorListResponse'
                    description: API rate limit exceeded.
                '500':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorListResponse'
                    description: Something is broken.
                '502':
                    description: The cloud is down or being upgraded.
                '503':
                    description: The cloud is up but overloaded with requests.
                '504':
                    description: The cloud is up but there is a failure somewhere causing the request to fail.
            security:
                -
                    bearerAuth: []
            operationId: get_users_profile
            summary: Get Employee Profile
            x-openapi-router-controller: domains.employees.controllers.get.controller
    '/employees/{id}':
        get:
            tags:
                - Employees
            parameters:
                -
                    style: simple
                    explode: false
                    name: id
                    schema:
                        type: string
                    in: path
                    required: true
            responses:
                '200':
                    description: Successfully retrieved Employee
                '429':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorListResponse'
                    description: API rate limit exceeded.
                '500':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorListResponse'
                    description: Something is broken.
                '502':
                    description: The cloud is down or being upgraded.
                '503':
                    description: The cloud is up but overloaded with requests.
                '504':
                    description: The cloud is up but there is a failure somewhere causing the request to fail.
            security:
                -
                    bearerAuth: []
            operationId: get_employee
            summary: Get Employee
            x-openapi-router-controller: domains.employees.controllers.get.controller
        delete:
            tags:
                - Employees
            parameters:
                -
                    style: simple
                    explode: false
                    name: id
                    schema:
                        type: string
                    in: path
                    required: true
            responses:
                '204':
                    description: Employee deleted
                '429':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorListResponse'
                    description: API rate limit exceeded.
                '500':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorListResponse'
                    description: Something is broken.
                '502':
                    description: The cloud is down or being upgraded.
                '503':
                    description: The cloud is up but overloaded with requests.
                '504':
                    description: The cloud is up but there is a failure somewhere causing the request to fail.
            security:
                -
                    bearerAuth: []
            operationId: delete_employee
            summary: Delete Employee
            x-openapi-router-controller: domains.employees.controllers.delete.controller
        patch:
            requestBody:
                content:
                    application/json:
                        schema:
                            type: array
                            items:
                                $ref: '#/components/schemas/PatchDocument'
                required: true
            tags:
                - Employees
            parameters:
                -
                    style: simple
                    explode: false
                    name: id
                    schema:
                        type: string
                    in: path
                    required: true
            responses:
                '204':
                    description: Employee patched
                '429':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorListResponse'
                    description: API rate limit exceeded.
                '500':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorListResponse'
                    description: Something is broken.
                '502':
                    description: The cloud is down or being upgraded.
                '503':
                    description: The cloud is up but overloaded with requests.
                '504':
                    description: The cloud is up but there is a failure somewhere causing the request to fail.
            security:
                -
                    bearerAuth: []
            operationId: patch_employee
            summary: Patch Employee
            x-openapi-router-controller: domains.employees.controllers.patch.controller
    /employees/login:
        post:
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UserLoginRequest'
            tags:
                - Employees
            responses:
                '200':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/UserLoginResponse'
                    description: Login successful
                '429':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorListResponse'
                    description: API rate limit exceeded.
                '500':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorListResponse'
                    description: Something is broken.
                '502':
                    description: The cloud is down or being upgraded.
                '503':
                    description: The cloud is up but overloaded with requests.
                '504':
                    description: The cloud is up but there is a failure somewhere causing the request to fail.
            operationId: authenticate_user
            summary: Login User
            x-openapi-router-controller: domains.employees.controllers.post.controller
    /user/signup:
        post:
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/User'
                x-body-name: employee
            tags:
                - Employees
            responses:
                '200':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/UserSignUpResponse'
                    description: User signup successful!
                '429':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorListResponse'
                    description: API rate limit exceeded.
                '500':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorListResponse'
                    description: Something is broken.
                '502':
                    description: The cloud is down or being upgraded.
                '503':
                    description: The cloud is up but overloaded with requests.
                '504':
                    description: The cloud is up but there is a failure somewhere causing the request to fail.
            security:
                -
                    bearerAuth: []
            operationId: signup_user
            summary: Signup Employee
            x-openapi-router-controller: domains.employees.controllers.post.controller
...
components:
    schemas:
...
        Department:
            title: Department
            description: Subdivisions of companies.
            required:
                - department_name
            type: object
            properties:
                id:
                    description: The id of the department.
                    type: integer
                department_name:
                    description: Block of text containing the name of the department.
                    type: string
                manager_id:
                    description: An id representing the ID of the department manager.
                    type: integer
                company_id:
                    description: The id representing the company the department is in.
                    type: integer
            example: '{''department_name'': ''Example'', ''manager_id'': 1, ''company_id'': 2 }'
...
        ErrorListResponse:
            title: Root Type for ErrorListResponse
            description: A list of error messages.
            required:
                - errors
            type: object
            properties:
                errors:
                    description: List of all error messages.
                    type: array
                    items:
                        title: ErrorMessage
                        type: object
            example: >-
                {'detail': 'The server encountered an unexpected error. Please try again later.','status':
                500,'suggestion': 'Please wait 5 minutes and re-attempt your request.'}
...
        ListResponse:
            title: ListResponse
            description: A response that contains data.
            required:
                - data
                - message
                - status
            type: object
            properties:
                data:
                    title: data
                    type: array
                    items:
                        $ref: '#/components/schemas/Department'
                message:
                    title: message
                    type: string
                status:
                    title: status
                    type: integer
            example:
                message: Success.
                status: 200
                data:
                    -
                        zipCode: zipCode
                        phone: phone
                        city: city
                        name: name
                        about: about
                        addressStreet: addressStreet
                        id: id
                        state: state
                    -
                        zipCode: zipCode
                        phone: phone
                        city: city
                        name: name
                        about: about
                        addressStreet: addressStreet
                        id: id
                        state: state
...
        PatchDocument:
            title: PatchDocument
            description: A JSONPatch document as defined by RFC 6902.
            required:
                - op
                - path
            type: object
            properties:
                op:
                    description: The operation to be performed
                    enum:
                        - add
                        - remove
                        - replace
                        - move
                        - copy
                        - test
                    type: string
                path:
                    description: A JSON-Pointer
                    type: string
                value:
                    description: The value to be used within the operations.
                    type: string
                fromPath:
                    description: A string containing a JSON Pointer value.
                    type: string
            example: |-
                {
                    "op": "replace",
                    "path": "/name",
                    "value": "Sample3"
                }
...
        User:
            title: User
            description: A single user.
            required:
                - email
            type: object
            properties:
                id:
                    description: A unique identifier for a user.
                    type: integer
                email:
                    description: email
                    type: string
                first_name:
                    description: The user's first name.
                    type: string
                last_name:
                    description: The user's last name.
                    type: string
                pay_rate:
                    description: The user's pay rate.
                    type: number
                department_id:
                    description: The user's department id.
                    type: integer
                role_id:
                    description: The user's role id.
                    type: integer
            example: |-
                {
                    "name": "User1",
                    "description": "User1 description"
                }
        UserLoginRequest:
            title: UserLoginRequest
            description: Request for user login.
            required:
                - email
                - password
            type: object
            properties:
                email:
                    title: email
                    description: The users email
                    type: string
                password:
                    title: password
                    description: The users password
                    type: string
            example:
                email: email
                password: password
        UserLoginResponse:
            title: UserLoginResponse
            description: Response on user login.
            required:
                - accessToken
            type: object
            properties:
                accessToken:
                    title: accessToken
                    description: accessToken that can be passed as authorization for endpoints
                    type: string
            example:
                accessToken: accessToken
        UserSignUpResponse:
            title: UserSignUpResponse
            description: Response to user signup
            required:
                - id
            type: object
            properties:
                id:
                    title: id
                    type: string
            example:
                id: id
        UserValidationList:
            title: UserValidationList
            description: A list of user validations.
            required:
                - employee_validations
            type: object
            properties:
                employee_validations:
                    title: data
                    type: array
                    items:
                        $ref: '#/components/schemas/UserValidation'
            example:
                message: Success.
                status: 200
                data:
                    -
                        zipCode: zipCode
                        phone: phone
                        city: city
                        name: name
                        about: about
                        addressStreet: addressStreet
                        id: id
                        state: state
                    -
                        zipCode: zipCode
                        phone: phone
                        city: city
                        name: name
                        about: about
                        addressStreet: addressStreet
                        id: id
                        state: state
        UserValidation:
            title: UserValidation
            description: A user validation request object.
            required:
                - first_name
                - last_name
                - company_name
                - department_name
            type: object
            properties:
                email:
                    title: email
                    type: string
                first_name:
                    title: first_name
                    type: string
                last_name:
                    title: last_name
                    type: string
                company_name:
                    title: company_name
                    type: string
                department_name:
                    title: department_name
                    type: string
                pay_rate:
                    title: pay_rate
                    type: number
                company_id:
                    title: company_id
                    type: integer
                department_id:
                    title: department_id
                    type: integer
                is_email_valid:
                    title: is_name_valid
                    type: boolean
                is_company_valid:
                    title: is_company_valid
                    type: boolean
                is_department_valid:
                    title: is_department_valid
                    type: boolean
                is_pay_rate_valid:
                    title: is_pay_rate_valid
                    type: boolean
                is_employee_valid:
                    title: is_employee_valid
                    type: boolean
...
        UserListResponse:
            title: ListResponse
            description: A response that contains data.
            required:
                - data
                - message
                - status
            type: object
            properties:
                data:
                    title: data
                    type: array
                    items:
                        $ref: '#/components/schemas/User'
                message:
                    title: message
                    type: string
                status:
                    title: status
                    type: integer
            example:
                message: Success.
                status: 200
                data:
                    -
                        zipCode: zipCode
                        phone: phone
                        city: city
                        name: name
                        about: about
                        addressStreet: addressStreet
                        id: id
                        state: state
                    -
                        zipCode: zipCode
                        phone: phone
                        city: city
                        name: name
                        about: about
                        addressStreet: addressStreet
                        id: id
                        state: state
...
    securitySchemes:
        bearerAuth:
            scheme: bearer
            bearerFormat: JWT
            type: http
            x-bearerInfoFunc: crosscutting.auth.authentication.authenticate_user
...